Update docker-compose.yaml
It is recommended to include the Email-related ENV configurations in the worker as well, otherwise the worker cannot access the necessary variables for sending emails when executing the email action. Having these SMTP configurations set only in the API is ineffective. Additionally, the CONSOLE_WEB_URL should also be set in the worker's ENV to ensure that the login URL in the emails received by users is correct.