feat(api): Add `EnumText` column type for SQLAlchemy
Introduce `EnumText`, a custom column type for SQLAlchemy designed
to work seamlessly with enumeration classes based on `StrEnum`.
This type stores enum members as `VARCHAR` in the database and
automatically handles conversion between the enumeration type and
its string representation during reads and writes.

Additionally, it validates that the stored and retrieved values
are valid members of the associated enumeration class.